LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/clients-matchall-include-uncontrolled.https-expected.txt

CONSOLE MESSAGE: line 2659: Error: assert_array_equals: property 1, expected false but got true


Harness Error (FAIL), message = Error: assert_array_equals: property 1, expected false but got true

TIMEOUT Verify matchAll() respect includeUncontrolled Test timed out
